# docs/swagger.yaml
openapi: 3.0.3
info:
  title: SIU-2 Backend API
  version: 1.0.0
  description: Documentación OpenAPI 3.0 generada exclusivamente a partir de los routers definidos en routes/.
servers:
  - url: http://localhost:5000/api
    description: Servidor local
...
security:
  - CookieAuth: []
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token_cookie
      description: Cookie con el JWT de sesión.
  schemas:
    ErrorItem:
      type: object
      required:
        - code
        - message
        - level
      properties:
        code:
          type: string
        message:
          type: string
        level:
          type: string
        description:
          type: string
    ErrorResponse:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorItem'
...
    Materia:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        codigo:
          type: string
          nullable: true
    MateriaCreateRequest:
      type: object
      required:
        - nombre
      properties:
        nombre:
          type: string
        codigo:
          type: string
          nullable: true
    MateriaUpdateRequest:
      type: object
      required:
        - nombre
        - codigo
      properties:
        nombre:
          type: string
        codigo:
          type: string
          nullable: true
    Curso:
      type: object
      properties:
        id:
          type: integer
        materia_id:
          type: integer
        nombre:
          type: string
        anio:
          type: integer
        cuatrimestre:
          type: integer
...
    Nota:
      type: object
      properties:
        id:
          type: integer
        evaluacion_id:
          type: integer
        alumno_id:
          type: integer
          nullable: true
        equipo_id:
          type: integer
          nullable: true
        nota:
          type: number
        observaciones:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    NotaIndividualCreateRequest:
      type: object
      required:
        - evaluacion_id
        - alumno_id
        - nota
      properties:
        evaluacion_id:
          type: integer
        alumno_id:
          type: integer
        nota:
          type: number
        observaciones:
          type: string
          nullable: true
    NotaGrupalCreateRequest:
      type: object
      required:
        - evaluacion_id
        - equipo_id
        - nota
      properties:
        evaluacion_id:
          type: integer
        equipo_id:
          type: integer
        nota:
          type: number
        observaciones:
          type: string
          nullable: true
    NotaCreateRequest:
      oneOf:
        - $ref: '#/components/schemas/NotaIndividualCreateRequest'
        - $ref: '#/components/schemas/NotaGrupalCreateRequest'
    NotaUpdateRequest:
      type: object
      required:
        - nota
      properties:
        nota:
          type: number
  responses:
    BadRequest:
      description: Solicitud inválida o datos de entrada no válidos.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Falta la cookie de autenticación, el token es inválido o expiró.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: El usuario autenticado no posee el rol requerido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: El recurso solicitado no existe.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: El recurso ya existe o viola una restricción de unicidad.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Error inesperado del servidor.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
...
  /materias:
    get:
      tags: [Materias]
      summary: Listar materias
      parameters:
        - name: nombre
          in: query
          schema:
            type: string
        - name: codigo
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Lista paginada de materias.
          content:
            application/json:
              schema:
                type: object
                required: [materias, total, page, page_size]
                properties:
                  materias:
                    type: array
                    items:
                      $ref: '#/components/schemas/Materia'
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
        '204':
          description: Sin resultados.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Materias]
      summary: Crear una materia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MateriaCreateRequest'
      responses:
        '201':
          description: Materia creada correctamente.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                required: [message, materia]
                properties:
                  message:
                    type: string
                  materia:
                    $ref: '#/components/schemas/Materia'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /materias/{materia_id}:
    parameters:
      - name: materia_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Materias]
      summary: Obtener una materia por ID
      responses:
        '200':
          description: Materia encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Materia'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Materias]
      summary: Reemplazar una materia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MateriaUpdateRequest'
      responses:
        '200':
          description: Materia actualizada correctamente.
          content:
            application/json:
              schema:
                type: object
                required: [message, materia]
                properties:
                  message:
                    type: string
                  materia:
                    $ref: '#/components/schemas/Materia'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Materias]
      summary: Eliminar una materia
      responses:
        '204':
          description: Materia eliminada correctamente.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /materias/{materia_id}/cursos:
    get:
      tags: [Materias]
      summary: Listar cursos de una materia
      parameters:
        - name: materia_id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Lista paginada de cursos asociados a la materia.
          content:
            application/json:
              schema:
                type: object
                required: [cursos, total, page, page_size]
                properties:
                  cursos:
                    type: array
                    items:
                      $ref: '#/components/schemas/Curso'
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
        '204':
          description: Sin resultados.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /notas:
    get:
      tags: [Notas]
      summary: Listar notas
      parameters:
        - name: evaluacion_id
          in: query
          schema:
            type: integer
        - name: alumno_id
          in: query
          schema:
            type: integer
        - name: equipo_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Lista de notas.
          content:
            application/json:
              schema:
                type: object
                required: [notas, total]
                properties:
                  notas:
                    type: array
                    items:
                      $ref: '#/components/schemas/Nota'
                  total:
                    type: integer
        '204':
          description: Sin resultados.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Notas]
      summary: Crear una nota
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotaCreateRequest'
      responses:
        '201':
          description: Nota creada correctamente.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                required: [message, nota]
                properties:
                  message:
                    type: string
                  nota:
                    $ref: '#/components/schemas/Nota'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /notas/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Notas]
      summary: Obtener una nota por ID
      responses:
        '200':
          description: Nota encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Nota'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags: [Notas]
      summary: Actualizar parcialmente una nota
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotaUpdateRequest'
      responses:
        '204':
          description: Nota actualizada correctamente.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: La nota no existe.
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error:
                    type: string
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Notas]
      summary: Eliminar una nota
      responses:
        '204':
          description: Nota eliminada correctamente.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...